requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.9.0            # via dm-haiku, jax, jaxlib
dm-haiku==0.0.1b0         # via pandas_toolkit (setup.py)
dm-tree==0.1.5            # via dm-haiku
jax==0.1.68               # via pandas_toolkit (setup.py)
jaxlib==0.1.47            # via pandas_toolkit (setup.py)
joblib==0.15.1            # via scikit-learn
numpy==1.18.4             # via dm-haiku, jax, jaxlib, opt-einsum, pandas, scikit-learn, scipy
opt-einsum==3.2.1         # via jax
pandas==1.0.3             # via pandas_toolkit (setup.py)
python-dateutil==2.8.1    # via pandas
pytz==2020.1              # via pandas
scikit-learn==0.23.1      # via pandas_toolkit (setup.py)
scipy==1.4.1              # via jaxlib, scikit-learn
six==1.15.0               # via absl-py, dm-haiku, dm-tree, python-dateutil
tabulate==0.8.7           # via dm-haiku
threadpoolctl==2.0.0      # via scikit-learn
wrapt==1.12.1             # via dm-haiku
